Split component for determining which extensions are needed for a project
